pyproject.toml
[project]
name = "ai-dev-environment"
version = "0.8.0"
description = "AI-powered development environment with multi-agent code documentation generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "langgraph>=0.2.0", # Multi-agent orchestration
    "pydantic>=2.0.0", # Data validation
    "pyyaml>=6.0.0", # YAML configuration
    "langchain>=0.3.0", # LangChain core
    "langchain-openai>=0.2.0", # OpenAI and OpenAI-compatible models (Ollama, OpenRouter, Together AI)
    "python-dotenv>=1.0.0", # Environment variable management
    # CLI and UI
    "click>=8.1.0", # CLI framework
    "rich>=13.0.0", # Terminal formatting
    "loguru>=0.7.0", # Logging
    # Testing
    "pytest>=7.4.0", # Test framework
    "pytest-cov>=4.1.0", # Coverage reporting
    "tree-sitter>=0.25.2",
    "tree-sitter-c-sharp>=0.23.1",
]

[project.scripts]
docugen = "src.modules.docugen.workflows.parser_led.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.uv]
package = true  # Enable entry point installation with uv sync
